Handle saving changes made to todo item
Add click event listener to todoItemModal with handler
handleTodoSaveChanges. This updates the todo item in storage with
any changes to it's properties, closes to todo item modal and
re-renders the space list to show any updates.
populateTodoItemModal method was changed to take todoID as a
parameter and set it as a dataset attribute on the modal. This supports
save the changes of the todo item by using it's id.