Update and create logic to render either pending or completed todos
Remove the completed space as part of initialization logic, rather
all completed todos with status 'commpleted' will be displayed as a
view rather than stored seperately in storage.
Update the getTodoList function to getPendingTodoList, to be used for
standard space lists of pending tasks. Create a new function
getCompletedTodoList to support displaying the completed list.
Create a Display function to render the list of completed todos.